Update Laravel & Use integer for the price

+++ database/migrations/2024_08_03_154028_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('slug');
             $table->json('images')->nullable();
             $table->longText('description')->nullable();
-            $table->decimal('price', 10, 2);
+            $table->unsignedBigInteger('price');
             $table->boolean('is_active')->default(true);
             $table->boolean('is_featured')->default(false);
             $table->boolean('in_stock')->default(true);
